Refactor service functions outside of api routes
