+++ src/app/api/school/students/route.ts
@@ -1,31 +1,14 @@
 import { NextResponse } from "next/server";
-import { createClient } from "@supabase/supabase-js";
 import { requireOrgAdmin } from "@/features/school/auth/org-admin";
+import { getStudents } from "@/features/school/services/students";
 
 export async function GET() {
   const auth = await requireOrgAdmin();
   if (auth instanceof NextResponse) return auth;
-  const { orgId } = auth;
-
-  const supabase = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.SUPABASE_SERVICE_ROLE_KEY!,
-  );
-
-  const { data: students, error } = await supabase
-    .from("profiles")
-    .select(
-      "user_id, first_name, last_name, title, education, city, country, is_technical, created_at, deleted_at",
-    )
-    .eq("organization_id", orgId)
-    .order("created_at", { ascending: false });
-
-  if (error) {
-    return NextResponse.json(
-      { error: "Failed to fetch students" },
-      { status: 500 },
-    );
+  try {
+    const students = await getStudents(auth.orgId);
+    return NextResponse.json({ students });
+  } catch {
+    return NextResponse.json({ error: "Failed to fetch students" }, { status: 500 });
   }
-
-  return NextResponse.json({ students: students ?? [] });
 }
